Organize src folder
