+++ src/EVN.h
@@ -1,24 +1,24 @@
 #ifndef EVN_h
 #define EVN_h
 
-#include "PIDController.h"
+#include "helper/PIDController.h"
 
 #include "EVNAlpha.h"
-#include "EVNMotor.h"
-#include "EVNServo.h"               //Geekservo (SERVO)
-#include "EVNColourSensor.h"        //TCS34725
-#include "EVNDistanceSensor.h"      //VL53L0X
-#include "EVNCompassSensor.h"       //HMC5883L
-// #include "EVNGyroSensor.h"       //MPU-6500
-#include "EVNEnvSensor.h"           //BME-280
-// #include "EVNGestureSensor.h"    //APDS-9960
-#include "EVNTouchArray.h"          //MPR121
-#include "EVNRGBLED.h"              //WS2812B (SERVO)
-// #include "EVNMatrixLED.h"        //HT16K33
-// #include "EVNSevenSegmentLED.h"  //HT16K33
-#include "EVNDisplay.h"             //SSD1306
-#include "EVNBluetooth.h"           //HC-05 (UART)
-// #include "EVNAnalogMux.h"        //ADS1115
+#include "actuators/EVNMotor.h"
+#include "actuators/EVNServo.h"               //Geekservo (SERVO)
+#include "sensors/EVNColourSensor.h"        //TCS34725
+#include "sensors/EVNDistanceSensor.h"      //VL53L0X
+#include "sensors/EVNCompassSensor.h"       //HMC5883L
+// #include "sensors/EVNGyroSensor.h"       //MPU-6500
+#include "sensors/EVNEnvSensor.h"           //BME-280
+// #include "sensors/EVNGestureSensor.h"    //APDS-9960
+#include "sensors/EVNTouchArray.h"          //MPR121
+#include "displays/EVNRGBLED.h"              //WS2812B (SERVO)
+// #include "displays/EVNMatrixLED.h"        //HT16K33
+// #include "displays/EVNSevenSegmentLED.h"  //HT16K33
+#include "displays/EVNDisplay.h"             //SSD1306
+#include "others/EVNBluetooth.h"           //HC-05 (UART)
+// #include "others/EVNAnalogMux.h"        //ADS1115
 
 // #include "non-standard/EVN_RCWL1601.h"
