+++ src/EVNAlpha.h
@@ -1,8 +1,8 @@
 #ifndef EVNAlpha_h
 #define EVNAlpha_h
 
-#include "EVNButton.h"
-#include "EVNPortSelector.h"
+#include "helper/EVNButton.h"
+#include "helper/EVNPortSelector.h"
 #include "evn_alpha_pins.h"
 #include <Arduino.h>
