+++ src/helper/EVNButton.h
@@ -3,7 +3,7 @@
 
 #include <Arduino.h>
 #include "EVNISRTimer.h"
-#include "evn_alpha_pins.h"
+#include "../evn_alpha_pins.h"
 
 #define BUTTON_DISABLE		0
 #define BUTTON_TOGGLE		1
