+++ src/helper/EVNI2CDevice.h
@@ -1,14 +1,14 @@
-#ifndef EVNSensor_h
-#define EVNSensor_h
+#ifndef EVNI2CDevice_h
+#define EVNI2CDevice_h
 
 #include <Arduino.h>
 #include <Wire.h>
-#include "EVNAlpha.h"
+#include "../EVNAlpha.h"
 
-class EVNSensor {
+class EVNI2CDevice {
 
 public:
-    EVNSensor(uint8_t port)
+    EVNI2CDevice(uint8_t port)
     {
         _port = constrain(port, 1, 16);
         if (_port <= 8)
